Initial commit for refactoring
This initial commit denotes the start of an attempt to go through a set
of massive refactoring. The rational behind the refactoring is
maintainability and decoupling of the various modules that now hide
within one big package. The things that require decoupling are PSL
related code from IO and vscode extension related code.

The aim is to use the npm workspaces feature to still maintain the
pieces of code in one big repository, but to create a workspace per
submodule of the whole extension ecosystem. Modules should also be able
to be released and used independently from the extension. Examples of
these modules are:

- vscode-psl: Still the main `src` dir with all the extension related
  code.
- psl-parser: Module that can tokenize and parse PSL source code
- psl-lint: Module that can use parsed code and run static code scans
  over the results
- psl-lint-cli: A CLI to the psl-lint features
- psl-debug: Module for debugging M or PSL code (under development in a
  branch that is not integrated in develop/master)

There might be more logical modules (Profile networking for MTM and PBS
related code comes to mind), which might be introduced at a later time.