Add empty workspace for psl-parser
This commit introduces a new empty workspace for the psl-parser
submodule. This change required some split of configuration to make
things reusable. The main tsconfig.json file that is used prior to this
commit is split in two: A common part which also can be used by the
submodules and a specific config that extends the common config.

One of the changes in the typescript configuration is the move to
composite projects. The modules will use reference projects for faster
build times. This requires the projects to be composite. Build info is
tracked in tsconfig.tsbuildinfo files, which are now ignored from git,
and added to the cleanup scripts.

The psl-parser module has typescript as dependency, and during the
installation of this dependency it is noticed that the typescript
version that is currently used is somewhat outdated. Updated the
typescript dependency in the main project as well to be in sync with the
one of psl-parser. This did cause a big change on the package-lock.json
file, as this file also changed from version 1 to 2.