Copy parser files to the psl-parser module
The parser related files that should be in a dedicates psl-parser
submodule are now copied and slightly altered. This psl-parser module
should not rely on the `fs` module, so the `parseFile()` method has been
exluded.

The `index.ts` file now just exposes all the exports of the three other
modules in the parser submodule, but this should be reconsidered. This
is the easiest way for now to easily adjust the main extension code to
work with the parser project in the upcoming commits, but it is better
to only expose the parse methods and the interfaces etc. that represent
the results of a parsed piece of PSL code.

This commit also already cleans up the config that was used for the
original psl-parser npm package. Because this change will definitely
break any api on the currently released version 0.0.3, the version of
the new submodule is changed to 1.0.0-dev.