Use new psl-parser submodule
This commit removes the three modules that were moved to the psl-parser
submodule, and changed all the other modules that depended on these
three modules in the main src directory.

Exported the `getLineAfter()` method in the psl-parser submodule as
well. This probably needs to be a method on some interface instead of an
exported function. I can imagine that get line after is usefull for
methods to for an "inser method" function. Currently it is mainly used
to get the line after the method signature to insert a DOC block, or
assert that one is there.