Move parser tests to psl-parser module
Moved the parser, statementParser and tokenizer tests to the psl-package
module. Added a `jest.config.ts` file for the Jest configuration. Prefer
to keep this config in a dedicated file to not blow up the package.json
with a lot of config, and prefer the TypeScript form to keep the project
clean from js files.