Refactor test setup main module
Refactored the test setup in the main module to be in line with the
psl-parser module. This includes a rename of all the test modules from
`*-test.ts` to `*.test.ts`. This way, the default ts-jest preset can be
used to run Jest tests that are written in TypeScript. It also includes
removing the old config from the package.json file, which also included
some redundant config.