Move all PSL linter code to psl-lint submodule
Move all the code that is responsible for linting parsed documents to
the psl-lint module. Moved all the code that implements specific rules
to a rules subdirectory.

Did not move the activate.ts, config.ts and cli/ from pslLint to the
submodule, as those belong to either the main module (activate.ts), or
to the CLI (config.ts and cli/).