Fix psl-lint CLI dependencies in actions
The github actions where not properly installing the dependencies of
psl-lint. Also added `lib` to the excluded dirs of the psl-lint tsconfig
file, as recompiling resulted in a lot of errors because the tsc command
was considering the output files as input in the second run of npm run
compile-lint.