Update json5 to 2.2.3 and the version to 1.13.4
Update the version of the json5 dependency to 2.2.3 and update the
version of this project to 1.13.4. Note that this commit also added the
changelog entry for 1.12.2. The current branch that is used for
development of the 1.13 version of this app has massive changes compared
to 1.12. That is why the changes are applied in two different branches.