Update scrips in package.json
Update the scripts in the package.json with the new commands to
bundle and build the extension. Also update the launch and task
JSON files for debugging the plugin.

Reverted some of the ts extensions from the imports. The three
dependend packages do not distribute the ts files, therefore these
imports require a js extension still. These dependencies should be
updated to export all that is requires from the index or when
bundled this would work probably as well.